Struct piston::cam::Camera

pub struct Camera<T = f32> {
    pub position: [T, ..3],
    pub up: [T, ..3],
    pub right: [T, ..3],
    pub forward: [T, ..3],
}

Models a camera with position and directions.

Fields

position

The camera position.

up

The up direction.

right

The right direction.

forward

The forward direction.

Methods

impl<T: Copy + FloatMath + Float> Camera<T>

fn new(x: T, y: T, z: T) -> Camera<T>

Constructs a new camera.

Places the camera at [x, y, z], looking towards pozitive z.

fn orthogonal(&self) -> [[T, ..4], ..4]

Computes an orthogonal matrix for the camera.

This matrix can be used to transform coordinates to the screen.

fn look_at(&mut self, x: T, y: T, z: T)

Orients the camera to look at a point.

fn set_yaw_pitch(&mut self, yaw: T, pitch: T)

Sets yaw and pitch angle of camera in radians.