Module piston::event

Modules

generic_event
Trait for generic events

Structs

AfterRenderArgs
After render arguments.

EventId
Used to identify events arguments provided by traits.

IdleArgs
Idle arguments, such as expected idle time in seconds.

RenderArgs
Render arguments

UpdateArgs
Update arguments, such as delta time in seconds

WindowEvents
An event loop iterator

Enums

Event
Adds render and update events to input events

Constants

DEFAULT_MAX_FPS
The default maximum frames per second.

DEFAULT_UPS
The default updates per second.

Traits

AfterRenderEvent
After rendering and buffers are swapped.

EventMap
Methods required to map from consumed event to emitted event.

Events
A trait for create event iterator from window.

FocusEvent
When window gets or looses focus

GenericEvent
Implemented by all events

IdleEvent
When background tasks should be performed

MouseCursorEvent
The position of the mouse cursor

MouseRelativeEvent
The relative movement of mouse cursor

MouseScrollEvent
The scroll of the mouse wheel

PressEvent
The press of a button

ReleaseEvent
The release of a button

RenderEvent
When the next frame should be rendered

ResizeEvent
When the window is resized

TextEvent
When receiving text from user, such as typing a character

UpdateEvent
When the application state should be updated