Struct piston::event::event_loop::Events

pub struct Events<W, E> where W: Window, E: EventMap<W::Event> {
    // some fields omitted
}

An event loop iterator

Warning: Because the iterator polls events from the window back-end, it must be used on the same thread as the window back-end (usually main thread), unless the window back-end supports multi-thread event polling.

Example:

fn main() {
    let opengl = shader_version::opengl::OpenGL_3_2;
    let window = Sdl2Window::new(
        opengl,
        WindowSettings {
            title: "Example".to_string(),
            size: [500, 500],
            fullscreen: false,
            exit_on_esc: true,
            samples: 0,
        }
    )
    let ref mut gl = Gl::new();
    let window = RefCell::new(window);
    for e in Events::new(&window)
        .set(Ups(120))
        .set(MaxFps(60)) {
        use event::RenderEvent;
        e.render(|args| {
            // Set the viewport in window to render graphics.
            gl.viewport(0, 0, args.width as i32, args.height as i32);
            // Create graphics context with absolute coordinates.
            let c = Context::abs(args.width as f64, args.height as f64);
            // Do rendering here.
        });
    }
}

Methods

impl<W, E> Events<W, E> where W: Window, E: EventMap<W::Event>

fn new(window: Rc<RefCell<W>>) -> Events<W, E>

Creates a new event iterator with default UPS and FPS settings.

fn ups(self, frames: u64) -> Events<W, E>

The number of updates per second

This is the fixed update rate on average over time. If the event loop lags, it will try to catch up.

fn max_fps(self, frames: u64) -> Events<W, E>

The maximum number of frames per second

The frame rate can be lower because the next frame is always scheduled from the previous frame. This causes the frames to "slip" over time.

fn swap_buffers(self, enable: bool) -> Events<W, E>

Enable or disable automatic swapping of buffers.