Module piston::gfx::shade

Shaders Shader parameter handling.

Structs

DictionaryProgram
A shader program with dictionary of parameters

NamedCell
A named cell containing arbitrary value

ParamDictionary
A dictionary of parameters, meant to be shared between different programs

ParamDictionaryLink
An associated link structure for ParamDictionary that redirects program input to the relevant dictionary cell.

ParamValues
A borrowed mutable storage for shader parameter values.

UserProgram
A bundle that encapsulates a program and a custom user-provided structure containing the program parameters.

Type parameters:

Enums

ParameterError
An error type on either the parameter storage or the program side

Traits

Program
Encloses a shader program handle with its parameter

ShaderParam
Abstracts the shader parameter structure, generated by the shader_param attribute

ToUniform
Helper trait to transform base types into their corresponding uniforms

Type Definitions

EmptyProgram

TextureParam

VarBlock

VarTexture

VarUniform