Module piston::gfx::state

Fixed-function hardware state.

Configures the primitive assembly (PA), rasterizer, and output merger (OM) blocks.

Structs

Blend

BlendChannel

ColorMask

Depth
Depth test state.

Factor

Primitive
Primitive rasterization state. Note that GL allows different raster method to be used for front and back, while this abstraction does not.

Stencil
Complete stencil state, specifying how to handle the front and back side of a face.

StencilSide
Complete stencil state for a given side of a face.

Enums

BlendValue

Comparison
A pixel-wise comparison function.

CullMode
Which face, if any, to cull.

Equation

InverseFlag

OffsetType
How to offset vertices in screen space, if at all.

RasterMethod
How to rasterize a primitive.

StencilOp
Stencil mask operation.

WindingOrder
The winding order of a set of vertices.

Statics

pub static Alpha: ColorMask = [definition]

pub static Blue: ColorMask = [definition]

pub static Green: ColorMask = [definition]

pub static MaskAll: ColorMask = [definition]

pub static Red: ColorMask = [definition]

Type Definitions

LineWidth

OffsetFactor

OffsetUnits