Struct piston::gfx::ClearData

pub struct ClearData {
    pub color: Option<[f32, ..4]>,
    pub depth: Option<f32>,
    pub stencil: Option<u8>,
}

How to clear a frame.

Fields

color

If set, the color buffer of the frame will be cleared to this.

depth

If set, the depth buffer of the frame will be cleared to this.

stencil

If set, the stencil buffer of the frame will be cleared to this.