Struct piston::gfx::tex::SamplerInfo

pub struct SamplerInfo {
    pub filtering: FilterMethod,
    pub wrap_mode: (WrapMode, WrapMode, WrapMode),
    pub lod_bias: f32,
    pub lod_range: (f32, f32),
}

Specifies how to sample from a texture.

Fields

filtering

Filter method to use.

wrap_mode

Wrapping mode for each of the U, V, and W axis (S, T, and R in OpenGL speak)

lod_bias

This bias is added to every computed mipmap level (N + lod_bias). For example, if it would select mipmap level 2 and lod_bias is 1, it will use mipmap level 3.

lod_range

This range is used to clamp LOD level used for sampling

Methods

impl SamplerInfo

fn new(filtering: FilterMethod, wrap: WrapMode) -> SamplerInfo

Create a new sampler description with a given filter method and wrapping mode, using no LOD modifications.