Enum piston::gfx::BufferUsage

pub enum BufferUsage {
    UsageStatic,
    UsageDynamic,
    UsageStream,
}

A hint as to how this buffer will be used.

The nature of these hints make them very implementation specific. Different drivers on different hardware will handle them differently. Only careful profiling will tell which is the best to use for a specific buffer.

Variants

UsageStatic

Once uploaded, this buffer will rarely change, but will be read from often.

UsageDynamic

This buffer will be updated "frequently", and will be read from multiple times between updates.

UsageStream

This buffer always or almost always be updated after each read.