Enum piston::gfx::Slice

pub enum Slice {
    VertexSlice(PrimitiveType, u32, u32),
    IndexSlice8(PrimitiveType, BufferHandle<u8>, u32, u32),
    IndexSlice16(PrimitiveType, BufferHandle<u16>, u32, u32),
    IndexSlice32(PrimitiveType, BufferHandle<u32>, u32, u32),
}

Description of a subset of Mesh data to render. We provide a primitive type in a slice because it is how we interpret mesh contents. For example, we can have a Point typed vertex slice to do shape blending, while still rendereing it as an indexed TriangleList.

Variants

VertexSlice

Render vertex data directly from the Mesh's buffer, using only the vertices between the two endpoints.

IndexSlice8

The IndexSlice* buffer contains a list of indices into the Mesh data, so every vertex attribute does not need to be duplicated, only its position in the Mesh. For example, when drawing a square, two triangles are needed. Using only VertexSlice, one would need 6 separate vertices, 3 for each triangle. However, two of the vertices will be identical, wasting space for the duplicated attributes. Instead, the Mesh can store 4 vertices and an IndexSlice8 can be used instead.

IndexSlice16

As IndexSlice8 but with u16 indices

IndexSlice32

As IndexSlice8 but with u32 indices