Module piston::graphics

Modules

internal
Contains internal type aliases and traits.

interpolation
Interpolation algorithms.

modular_index
Helper functions for computing modular index safely.

triangulation
Methods for converting shapes into triangles.

vecmath
Various methods for computing with vectors.

Structs

Context
Drawing 2d context.

Traits

AddBevel
Implemented by contexts that can make a shape bevel.

AddBevelBorder
Implemented by contexts that can add round border.

AddBorder
Implemented by contexts that can add border.

AddColor
Implemented by contexts who can add color.

AddEllipse
Implemented by all contexts that can add ellipse.

AddImage
Implemented by contexts that can add image.

AddLine
Implemented by all contexts that can add rectangle.

AddPolygon
Implemented by contexts who can add polygon.

AddPolygons
Implemented by tweening contexts that can add polygons.

AddRectangle
Implemented by all contexts that can add rectangle.

AddRound
Implemented by contexts that can make a shape rounder.

AddRoundBorder
Implemented by contexts that can add round border.

AddSquareBorder
Implemented by contexts that can add square border.

AddTween
Implemented by contexts who can give an animated inbetweening context.

BackEnd
Implemented by all graphics back-ends. This trait uses default methods to simplify implementation.

Draw
Implemented by contexts that can draws something using a back-end.

ImageSize
Must be implemented by all images to be used with graphics back-end.

RelativeColor
Implemented by contexts that contains color.

RelativeRectangle
Should be implemented by contexts that have rectangle information.

RelativeSourceRectangle
Should be implemented by contexts that have source rectangle information.

RelativeTransform2d
Implemented by contexts that can transform.

View
Should be implemented by contexts that draws something relative to view.

Type Definitions

BevelBorderLineColorContext

BevelBorderLineContext

BevelRectangleBorderColorContext

BevelRectangleBorderContext

BevelRectangleColorContext

BevelRectangleContext

ColorContext

EllipseBorderColorContext

EllipseBorderContext

EllipseColorContext

EllipseContext

ImageColorContext

ImageContext

ImageRectangleColorContext

ImageRectangleContext

LerpTweenColorContext

LerpTweenContext

LerpTweenPolygonsColorContext

LerpTweenPolygonsContext

LineColorContext

LineContext

PolygonColorContext

PolygonContext

RectangleBorderColorContext

RectangleBorderContext

RectangleColorContext

RectangleContext

RoundBorderLineColorContext

RoundBorderLineContext

RoundRectangleBorderColorContext

RoundRectangleBorderContext

RoundRectangleColorContext

RoundRectangleContext

SquareBorderLineColorContext

SquareBorderLineContext