Module piston::graphics::internal

Contains internal type aliases and traits.

These do not affect the normal usage of the library. Using these might lead to breaking changes. Made public to clarify documentation.

Traits

CanColor
Implemented by contexts that can color.

CanRectangle
Implemented by contexts that can set rectangle information.

CanSourceRectangle
Implemented by contexts that can set source rectangle information.

CanTransform
Implemented by contexts that can transform.

CanViewTransform
Implemented by contexts that can view transform.

HasColor
Implemented by contexts that uses type U as current color.

HasRectangle
Implemented by contexts that uses type U as current rectangle.

HasSourceRectangle
Implemented by contexts that uses type U as current source rectangle.

HasTransform
Implemented by contexts that uses type U as current transform.

HasViewTransform
Implemented by contexts that uses type U as current view transform.

Type Definitions

Area

Color

ColorComponent

Line

Polygon

Polygons

Radius

Ray

Rectangle

SourceRectangle

Triangle

Vec2d

Width