Module piston::graphics::interpolation

Interpolation algorithms.

Interpolation is used in animation, to describe smooth shapes and to make transitions. Any object that fullfill certain mathematical properties can be interpolated. A common technique is using one ore more 'numbers' controlling the mixture of states. The choice of interpolation algorithm depends often on the circumstances where it used.

Functions

cub_bez
Performs cubic beziér interpolation. This is done by interpolation between two quadratic beziér. For more information, see:

lerp
Performs linear interpolation. A linear interpolation consists of two states 'a' and 'b'. The 't' variable is a factor between 0 and 1 that gives weight to 'a' or 'b'. When 't' is zero then 'a' has full weight. When 't' is one then 'b' has full weight.

lerp_2
Performs linear interpolation on array of size 3.

lerp_3
Performs linear interpolation on array of size 3.

lerp_4
Performs linear interpolation on array of size 4.

lerp_5
Performs linear interpolation on array of size 5.

quad_bez
Performs quadratic beziér interpolation. This is done by nesting linear interpolations. For more information, see: