Struct piston::graphics::Context

pub struct Context<S = (), C = ()> {
    pub view: [[f64, ..3], ..2],
    pub transform: [[f64, ..3], ..2],
    // some fields omitted
}

Drawing 2d context.

Fields

view

View transformation.

transform

Current transformation.

Methods

impl Context<(), ()>

fn new() -> Context<(), ()>

Creates a new drawing context.

fn abs(w: f64, h: f64) -> Context<(), ()>

Creates a new drawing context in absolute coordinates.

This function assumes the default coordinate system being centered with x axis pointing to the right and y axis pointing up.

Returns a drawing context with origin in the upper left corner and x axis pointing to the right and y axis pointing down.