Trait piston::graphics::RelativeSourceRectangle

pub trait RelativeSourceRectangle<T> {
    fn src_rect(&self, x: i32, y: i32, w: i32, h: i32) -> T;
    fn src_rel(&self, x: i32, y: i32) -> T;
    fn src_flip_h(&self) -> T;
    fn src_flip_v(&self) -> T;
    fn src_flip_hv(&self) -> T;
}

Should be implemented by contexts that have source rectangle information.

Required Methods

fn src_rect(&self, x: i32, y: i32, w: i32, h: i32) -> T

Adds a source rectangle.

fn src_rel(&self, x: i32, y: i32) -> T

Moves to a relative source rectangle using the current source rectangle as tile.

fn src_flip_h(&self) -> T

Flips the source rectangle horizontally.

fn src_flip_v(&self) -> T

Flips the source rectangle vertically.

fn src_flip_hv(&self) -> T

Flips the source rectangle horizontally and vertically.