Module piston::graphics::vecmath

Various methods for computing with vectors.

Functions

area
Computes area of a simple polygon.

area_centroid
Computes the area and centroid of a simple polygon.

centroid
Computes centroid of a simple polygon.

get_scale
Extract scale information from matrix.

hsv
Transform color with hue saturation and value.

identity
Create an identity matrix.

inside_triangle
Returns true if point is inside triangle.

least_separation_4
Returns the least separation out of four. Each seperation can be computed using separation function. The separation returned can be used to solve collision of rectangles.

line_side
Returns a number that tells which side it is relative to a line.

margin_rectangle
Shrinks a rectangle by a factor on all sides.

modular_offset
Computes modular offset safely for numbers.

orient
Orients x axis to look at point.

relative_rectangle
Computes a relative rectangle using the rectangle as a tile.

relative_source_rectangle
Computes a relative source rectangle using the source rectangle as a tile.

rotate_radians
Creates a rotation matrix.

row_mat2x3_mul
Multiplies two matrices.

scale
Create a scale matrix.

separation
Compute the shortest vector from point to ray. A ray stores starting point and directional vector.

shear
Create a shear matrix.

translate
Creates a translation matrix.

triangle_face
Returns true if triangle is clockwise.

Type Definitions

Matrix2d

Scalar

Vec2d