Module piston::image

Modules

gif
Decoding of GIF Images

imageops
Image Processing Functions

jpeg
Decoding and Encoding of JPEG Images

png
Decoding and Encoding of PNG Images

ppm
Encoding of portable pixmap Images

webp
Decoding of Webp Images

Structs

ImageBuf
An Image whose pixels are contained within a vector

Luma
A type to hold a grayscale pixel

LumaA
A type to hold a grayscale pixel with an alpha channel

MutPixels
Mutable pixel iterator

Pixels
Immutable pixel iterator

Rgb
A type to hold an RGB pixel

Rgba
A type to hold an RGB pixel with an alpha channel

SubImage
A View into another image

Enums

ColorType
An enumeration over supported color types and their bit depths

DynamicImage
A Dynamic Image

FilterType
Available Sampling Filters

ImageError
An enumeration of Image Errors

ImageFormat
An enumeration of supported image formats. Not all formats support both encoding and decoding.

Traits

GenericImage
A trait for manipulating images.

ImageDecoder
The trait that all decoders implement

MutableRefImage
A trait for images that allow providing mutable references to pixels.

Pixel
A trait that all pixels implement.

Functions

load
Create a new image from a Reader

load_from_memory
Create a new image from a byte slice

open
Open the image located at the path specified. The image's format is determined from the path's file extension.

Type Definitions

ImageResult