Module piston::image::imageops

Image Processing Functions

Enums

FilterType
Available Sampling Filters

Functions

blur
Performs a Gaussian blur on the supplied image. sigma is a measure of how much to blur by.

brighten
Brighten the supplied image value is the amount to brighten each pixel by. Negative values decrease the brightness and positive values increase it.

contrast
Adjust the contrast of the supplied image contrast is the amount to adjust the contrast by. Negative values decrease the contrast and positive values increase the contrast.

crop
Return a mutable view into an image

filter3x3
Perform a 3x3 box filter on the supplied image. kernel is an array of the filter weights of length 9.

flip_horizontal
Flip an image horizontally

flip_vertical
Flip an image vertically

grayscale
Convert the supplied image to grayscale

invert
Invert each pixel within the supplied image This function operates in place.

resize
Resize the supplied image to the specified dimensions nwidth and nheight are the new dimensions. filter is the sampling filter to use.

rotate180
Rotate an image 180 degrees clockwise.

rotate270
Rotate an image 270 degrees clockwise.

rotate90
Rotate an image 90 degrees clockwise.

unsharpen
Performs an unsharpen mask on the supplied image sigma is the amount to blur the image by. threshold is the threshold for the difference between see https://en.wikipedia.org/wiki/Unsharp_masking#Digital_unsharp_masking