Struct piston::image::png::PNGEncoder

pub struct PNGEncoder<W> {
    // some fields omitted
}

The representation of a PNG encoder

Methods

impl<W: Writer> PNGEncoder<W>

fn new(w: W) -> PNGEncoder<W>

Create a new encoder that writes its output to w

fn encode(&mut self, image: &[u8], width: u32, height: u32, c: ColorType) -> Result<(), IoError>

Encodes the image image that has dimensions width and height and ColorType c