Enum piston::shader_version::VersionRange

pub enum VersionRange<T> {
    VersionFromTo(T, T),
    VersionFrom(T),
}

Models knowledge about version range.

Variants

VersionFromTo

The version range is working from a version, up to but not included another version.

With other words, from first time it works until it breaks.

VersionFrom

The version range is working from a version, and possibly for all newer versions unless the context it is used implies otherwise.