Trait piston::EventLoop

pub trait EventLoop {

    pub fn get_event_settings(&self) -> EventSettings;

    pub fn set_event_settings(&mut self, settings: EventSettings);

    pub fn set_ups(&mut self, frames: u64) { ... }

    pub fn ups(self, frames: u64) -> Self { ... }

    pub fn set_ups_reset(&mut self, frames: u64) { ... }

    pub fn ups_reset(self, frames: u64) -> Self { ... }

    pub fn set_max_fps(&mut self, frames: u64) { ... }

    pub fn max_fps(self, frames: u64) -> Self { ... }

    pub fn set_swap_buffers(&mut self, enable: bool) { ... }

    pub fn swap_buffers(self, enable: bool) -> Self { ... }

    pub fn set_bench_mode(&mut self, enable: bool) { ... }

    pub fn bench_mode(self, enable: bool) -> Self { ... }

    pub fn set_lazy(&mut self, enable: bool) { ... }

    pub fn lazy(self, enable: bool) -> Self { ... }

}

Methods implemented for changing event loop settings.

Required methods

pub fn get_event_settings(&self) -> EventSettings

Returns event loop settings.

pub fn set_event_settings(&mut self, settings: EventSettings)

Sets event loop settings.

Loading content...
Provided methods

pub fn set_ups(&mut self, frames: u64)

The number of updates per second

This is the fixed update rate on average over time. If the event loop lags, it will try to catch up. When set to 0, update events are disabled.

pub fn ups(self, frames: u64) -> Self

The number of updates per second

This is the fixed update rate on average over time. If the event loop lags, it will try to catch up. When set to 0, update events are disabled.

pub fn set_ups_reset(&mut self, frames: u64)

The number of delayed updates before skipping them to catch up. When set to 0, it will always try to catch up.

pub fn ups_reset(self, frames: u64) -> Self

The number of delayed updates before skipping them to catch up. When set to 0, it will always try to catch up.

pub fn set_max_fps(&mut self, frames: u64)

The maximum number of frames per second

The frame rate can be lower because the next frame is always scheduled from the previous frame. This causes the frames to “slip” over time.

pub fn max_fps(self, frames: u64) -> Self

The maximum number of frames per second

The frame rate can be lower because the next frame is always scheduled from the previous frame. This causes the frames to “slip” over time.

pub fn set_swap_buffers(&mut self, enable: bool)

Enable or disable automatic swapping of buffers.

pub fn swap_buffers(self, enable: bool) -> Self

Enable or disable automatic swapping of buffers.

pub fn set_bench_mode(&mut self, enable: bool)

Enable or disable benchmark mode. When enabled, it will render and update without sleep and ignore input. Used to test performance by playing through as fast as possible. Requires lazy to be set to false.

pub fn bench_mode(self, enable: bool) -> Self

Enable or disable benchmark mode. When enabled, it will render and update without sleep and ignore input. Used to test performance by playing through as fast as possible. Requires lazy to be set to false.

pub fn set_lazy(&mut self, enable: bool)

Enable or disable rendering only when receiving input. When enabled, update events are disabled. Idle events are emitted while receiving input.

pub fn lazy(self, enable: bool) -> Self

Enable or disable rendering only when receiving input. When enabled, update events are disabled. Idle events are emitted while receiving input.

Loading content...