Enum piston::Event

pub enum Event {
    Input(Input, Option<u32>),
    Loop(Loop),
    Custom(EventId, Arc<dyn Any + Send + Sync, Global>, Option<u32>),
}

Models all events.

Variants

Input(Input, Option<u32>)

Input events.

Time stamp is ignored when comparing input events for equality and order.

Loop(Loop)

Events that commonly used by event loops.

Custom(EventId, Arc<dyn Any + Send + Sync, Global>, Option<u32>)

Custom event.

When comparing two custom events for equality, they always return false.

When comparing partial order of two custom events, the event ids are checked and if they are equal it returns None.

Time stamp is ignored both when comparing custom events for equality and order.